Fix:Tools:Ticket #541, integer overflow in geom.c
git-svn-id: http://svn.code.sf.net/p/navit/code/trunk/navit@2944 ffa7fe5e-494d-0410-b361-a75ebd5db220

+++ navit/maptool/geom.c
@@ -220,7 +220,7 @@ static int
 clip_line_code(struct coord *p1, struct coord *p2, struct rect *r)
 {
 	int code1,code2,ret=1;
-	int dx,dy;
+	long long dx,dy;
 	code1=clipcode(p1, r);
 	if (code1)
 		ret |= 2;
